requirements.txt
asgiref==3.6.0
attrs==22.2.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.4
django-rest==0.8.7
djangorestframework==3.14.0
exceptiongroup==1.1.0
idna==3.4
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
pytest==7.2.1
pytest-django==4.5.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.14
